Download Asset From Databrary.

download_asset() has been superseded in favor of download_video() and the more general download_session_asset().

download_asset(
  asset_id = 1,
  session_id = 9807,
  file_name = "test.mp4",
  target_dir = tempdir(),
  return_response = FALSE,
  vb = FALSE
)

Arguments

asset_id

Asset id for target file.

session_id

Slot/session number where target file is stored.

file_name

Name for downloaded file.

target_dir

Directory to save the downloaded file. Default is a temporary directory given by a call to tempdir().

return_response

A Boolean value.

vb

A Boolean value. If TRUE provides verbose output.

Value

A string that is the full file name of the asset.

Examples

# \donttest{
download_asset() # Download's 'numbers' file from volume 1.
#> [1] "/var/folders/n3/6wm9hw792y344j61fqh2rw94mcvkj6/T//Rtmpx4wSA3/9807-1-2024-02-09-0836-57.mp4"
download_asset(asset_id = 11643, session_id = 9825, file_name = "rdk.mp4") 
#> [1] "/var/folders/n3/6wm9hw792y344j61fqh2rw94mcvkj6/T//Rtmpx4wSA3/rdk.mp4"
# Downloads a display with a random dot kinematogram (RDK).
# }