Extracts the code definitions from a Datavyu (.opf) file.

Extracts the code definitions from a Datavyu (.opf) file.

extract_dv_code_defs(
  in_dir = ".",
  dv_fn = "db",
  out_dir = in_dir,
  out_fn = paste0(out_dir, "/", "codes.csv"),
  auto_write_over = TRUE,
  code_regex = "^([a-zA-Z_]+[0-9]*[a-zA-Z_]*[0-9]*)",
  code_vals_regex = "\\)-([a-zA-Z\\/]+)\\|",
  code_type_regex = "([a-zA-Z]+)$",
  vb = FALSE
)

Arguments

in_dir

Input directory. Defaults to '.'

dv_fn

Full Datavyu file name. Defaults to 'db'

out_dir

Output directory. Defaults to in_dir

out_fn

Full output file name.

auto_write_over

A Boolean value. If TRUE, new output file overwrites old.

code_regex

A string specifying the regular expression to extract a code.

code_vals_regex

A string specifying the regular expression to extract a code value.

code_type_regex

A string specifying the regular expression for the code type.

vb

A boolean value. If TRUE, provides verbose output.

Examples

extract_dv_code_defs()
#> Error in extract_dv_code_defs(): File db cannot be found.