Lists all the assets/files in a given volume.

Lists all the assets/files in a given volume.

Usage

list_assets_in_volume(vol_id = 1, vb = FALSE)

Arguments

vol_id

Target volume number.

vb

A boolean value.

Value

List of assets.

Examples

list_assets_in_volume() # Assets in volume 1
#>    asset_id asset_type_id
#> 1     22412             6
#> 2      9826          -800
#> 3      9828          -800
#> 4      9830          -800
#> 5      9832          -800
#> 6      9839             6
#> 7      9834          -800
#> 8      9836          -800
#> 9      9838          -800
#> 10    12039             6
#> 11    17737          -800
#> 12    22408             6
#> 13    17719          -800
#> 14    22407             6
#> 15    17723          -800
#> 16    17721          -800
#> 17    22410             6
#> 18    17725          -800
#> 19    22411             6
#> 20    17727          -800
#> 21    19919          -800
#> 22    22406             6
#> 23    25497             6
#> 24    25495             6
#> 25    25496             6
#> 26    27227             6
#> 27    26446             6
#> 28    28026          -800
#> 29    26449             6
#> 30    26448          -800
#> 31    26452             6
#> 32    28028          -800
#> 33   153109             2
#> 34   153108             2
#> 35   117035            16
#> 36      314             6
#> 37     9200             5
#> 38        1          -800
#> 39      315             6
#> 40    28047             6
#> 41    28051          -800
#> 42    28048             6
#> 43    28053          -800
#> 44    28055          -800
#> 45    28049             6
#> 46    28062          -800
#> 47    42470             6
#> 48    36978             6
#> 49    37128             6
#> 50    37138          -800
#> 51    36945             6
#> 52    37130          -800
#> 53    36946             6
#> 54    37136          -800
#> 55    43940             6
#> 56    38626             6
#> 57    38643          -800
#> 58    38627             6
#> 59    38591          -800
#> 60    38628             6
#> 61    38593          -800
#> 62    38629             6
#> 63    38595          -800
#> 64    43077             6
#> 65    44774          -800
#> 66    44544             6
#> 67    42740             6
#> 68    42755             6
#> 69    45413             6
#> 70    46104             6
#> 71    46300          -800
#> 72    60627             6
#> 73    60542          -800
#> 74    60628             6
#> 75    60544          -800
#> 76    60629             6
#> 77    60546          -800
#> 78    60630             6
#> 79    60550          -800
#> 80    61222          -800
#> 81    61224          -800
#> 82    65936             6
#> 83    66015          -800
#>                                                                             name
#> 1                                                                         Slides
#> 2                                                                   Introduction
#> 3                                                                 Databrary demo
#> 4                                                             Databrary 1.0 plan
#> 5                                                                        Datavyu
#> 6                                                                   04-07-Slides
#> 7                                                     Overview and Policy Update
#> 8                                                              Private Beta demo
#> 9                                           Positioning Databrary for the future
#> 10                                                                        slides
#> 11                                                                 Board Meeting
#> 12                                                                Welcome slides
#> 13                                                Best practices in video coding
#> 14                                        SLIDES: Best practices in video coding
#> 15                                                              Datavyu tutorial
#> 16                                                              Lunch discussion
#> 17                                                      SLIDES: Lunch discussion
#> 18                                                       Open video data sharing
#> 19                                               SLIDES: Open video data sharing
#> 20                                                                Databrary demo
#> 21                                                                         video
#> 22                                                                        slides
#> 23                                  Slides: Video reuse, management, and sharing
#> 24       Slides: Policies and best practices for sharing and managing video data
#> 25 Slides: Best practices in behavioral video coding and introduction to Datavyu
#> 26                                                                        Agenda
#> 27                                             SLIDES: Introduction to Databrary
#> 28                                                     Introduction to Databrary
#> 29           SLIDES: Policies & best practices for sharing & managing video data
#> 30                   Policies & best practices for sharing & managing video data
#> 31                SLIDES: Introduction to Datavyu: Video coding & best practices
#> 32                        Introduction to Datavyu: Video coding & best practices
#> 33                                                     volumes-citations-monthly
#> 34                                          databrary-institutions-investigators
#> 35                                                  counting (demo Datavyu file)
#> 36                                                        NIH Proposal Narrative
#> 37                                                                          logo
#> 38                                                         counting (demo video)
#> 39                                                        NSF Proposal Narrative
#> 40                                             SLIDES: Introduction to Databrary
#> 41                                                     Introduction to Databrary
#> 42           SLIDES: Policies & best practices for sharing & managing video data
#> 43          Policies & best practices for sharing & managing video data (Part 1)
#> 44          Policies & best practices for sharing & managing video data (Part 2)
#> 45                SLIDES: Introduction to Datavyu: Video coding & best practices
#> 46                        Introduction to Datavyu: Video coding & best practices
#> 47                                             Workshop attendees survey results
#> 48                             SLIDES: Datavyu & Best Practices for Coding Video
#> 49                                             SLIDES: Introduction to Databrary
#> 50                                                     Introduction to Databrary
#> 51                               SLIDES: Policies for Sharing Identifiable Video
#> 52                                       Policies for Sharing Identifiable Video
#> 53                                 SLIDES: Managing Data to Accelerate Discovery
#> 54                                         Managing Data to Accelerate Discovery
#> 55                                                                Survey results
#> 56                                             SLIDES: Introduction to Databrary
#> 57                                                     Introduction to Databrary
#> 58                               SLIDES: Policies for Sharing Identifiable Video
#> 59                                       Policies for Sharing Identifiable Video
#> 60                                          SLIDES: Managing Data with Databrary
#> 61                                                  Managing Data with Databrary
#> 62                             SLIDES: Datavyu & Best Practices for Coding Video
#> 63                                     Datavyu & Best Practices for Coding Video
#> 64                                                           ICIS Survey Results
#> 65                                                      Databrary Pre-conference
#> 66                                             SLIDES: Introduction to Databrary
#> 67                                       SLIDES: Policies for Sharing Video Data
#> 68                                          SLIDES: Managing Data with Databrary
#> 69                             SLIDES: Datavyu & Best Practices for Coding Video
#> 70                               SLIDES: Databrary Board Meeting - June 13, 2016
#> 71                                VIDEO: Databrary Board Meeting - June 13, 2016
#> 72                                             SLIDES: Introduction to Databrary
#> 73                                                     Introduction to Databrary
#> 74                                                  SLIDES: Policies for Sharing
#> 75                                               Policies for sharing video data
#> 76                                          SLIDES: Managing data with Databrary
#> 77                                                  Managing data with Databrary
#> 78                             SLIDES: Datavyu & best practices for coding video
#> 79                                     Datavyu & best Practices for coding video
#> 80                                                     Introduction to Databrary
#> 81                                                     Policies for sharing data
#> 82                                               PLAY workshop slides 2016-12-16
#> 83                                            PLAY workshop videocast 2016-12-16
#>    permission       size duration            segment
#> 1           5    4573426       NA               NULL
#> 2           5   88610655   335883          0, 335883
#> 3           5  917124852  4425483    335883, 4761366
#> 4           5  899912341  4277835   4761366, 9039201
#> 5           5  764340542  3107147  9039201, 12146348
#> 6           5    3013342       NA               NULL
#> 7           5 1301079971  4964011         0, 4964011
#> 8           5 1419151887  5043051  4964011, 10007062
#> 9           5  882417129  3821387 10007062, 13828449
#> 10          5   10192940       NA               NULL
#> 11          5 5316455384 10976384        0, 10976384
#> 12          5     304566       NA                  0
#> 13          5 1384301751  3514603         0, 3514603
#> 14          5    2924252       NA             528181
#> 15          5  574598972  1594688   3514560, 5109248
#> 16          5 1602298195  3977558   5111054, 9088612
#> 17          5    6899381       NA            5112000
#> 18          5  710384623  1920512  9086719, 11007231
#> 19          5    4158160       NA            9087000
#> 20          5 1079191541  2827904 11007188, 13835092
#> 21          5 5561649547 14939008        0, 14939008
#> 22          5   24014506       NA               5000
#> 23          5   19543511       NA                  0
#> 24          5    6947725       NA               1000
#> 25          5   11216184       NA               2000
#> 26          5      55248       NA                  0
#> 27          5    8502489       NA                100
#> 28          5  969574682  3708160   1061959, 4770119
#> 29          5   18074689       NA            4741737
#> 30          5 2059896165  5477483  4742737, 10220220
#> 31          5   16562761       NA           10220209
#> 32          5  717531830  2725723 10220220, 12945943
#> 33          5        417       NA               NULL
#> 34          5       1826       NA               NULL
#> 35          5       1030       NA               NULL
#> 36          5    3826718       NA               NULL
#> 37          5     104419       NA               NULL
#> 38          5     394499    40000               NULL
#> 39          5    4120930       NA               NULL
#> 40          5    9500786       NA                  0
#> 41          5  886710056  2794880         0, 2794880
#> 42          5   18069955       NA            2792054
#> 43          5  763350644  2387051   2795785, 5182836
#> 44          5 1044753106  2983616   5183754, 8167370
#> 45          5   17821836       NA            8150330
#> 46          5 1550955220  4835115  8168370, 13003485
#> 47          5     402918       NA               NULL
#> 48          5   16143550       NA               NULL
#> 49          5   22240020       NA                  0
#> 50          5 1190956598  3402326         0, 3402326
#> 51          5    2225737       NA            1800000
#> 52          5  824306229  2450882   1827192, 4278074
#> 53          5   13901204       NA            2400000
#> 54          5 1412799175  3904128   2451165, 6355293
#> 55          5     191095       NA               NULL
#> 56          5   17971722       NA                  0
#> 57          5  706313005  2760992         0, 2760992
#> 58          5    4265477       NA            2811000
#> 59          5  622855933  2293056   2814628, 5107684
#> 60          5   10040436       NA            5095000
#> 61          5 1035603796  3676480   5108605, 8785085
#> 62          5   11171502       NA            8784000
#> 63          5  857979902  2537494  8785992, 11323486
#> 64          5    1772791       NA               NULL
#> 65          5 5207302273  9728726         0, 9728726
#> 66          5   10732100       NA               2000
#> 67          5    4238406       NA            2185146
#> 68          5    7741745       NA            4401856
#> 69          5   15475455       NA            6480140
#> 70          5   19602134       NA               NULL
#> 71          5 4651981876 13129280        0, 13129280
#> 72          5   10732100       NA                  0
#> 73          5  607629550  2058198         0, 2058198
#> 74          5    4238406       NA            2056000
#> 75          5  795931994  2468587   2059089, 4527676
#> 76          5    7741745       NA            4524000
#> 77          5  627556299  2002624   4528588, 6531212
#> 78          5   15475455       NA            6527000
#> 79          5 1033985826  3372779   6532121, 9904900
#> 80          5  671824312  1939563         0, 1939563
#> 81          5  687136253  1908544   1940470, 3849014
#> 82          5  278432416       NA               NULL
#> 83          5 1332848036 24567644        0, 24567644
#>                   mimetype extension                asset_type transcodable
#> 1          application/pdf       pdf         Portable document           NA
#> 2                video/mp4       mp4              MPEG-4 video         -800
#> 3                video/mp4       mp4              MPEG-4 video         -800
#> 4                video/mp4       mp4              MPEG-4 video         -800
#> 5                video/mp4       mp4              MPEG-4 video         -800
#> 6          application/pdf       pdf         Portable document           NA
#> 7                video/mp4       mp4              MPEG-4 video         -800
#> 8                video/mp4       mp4              MPEG-4 video         -800
#> 9                video/mp4       mp4              MPEG-4 video         -800
#> 10         application/pdf       pdf         Portable document           NA
#> 11               video/mp4       mp4              MPEG-4 video         -800
#> 12         application/pdf       pdf         Portable document           NA
#> 13               video/mp4       mp4              MPEG-4 video         -800
#> 14         application/pdf       pdf         Portable document           NA
#> 15               video/mp4       mp4              MPEG-4 video         -800
#> 16               video/mp4       mp4              MPEG-4 video         -800
#> 17         application/pdf       pdf         Portable document           NA
#> 18               video/mp4       mp4              MPEG-4 video         -800
#> 19         application/pdf       pdf         Portable document           NA
#> 20               video/mp4       mp4              MPEG-4 video         -800
#> 21               video/mp4       mp4              MPEG-4 video         -800
#> 22         application/pdf       pdf         Portable document           NA
#> 23         application/pdf       pdf         Portable document           NA
#> 24         application/pdf       pdf         Portable document           NA
#> 25         application/pdf       pdf         Portable document           NA
#> 26         application/pdf       pdf         Portable document           NA
#> 27         application/pdf       pdf         Portable document           NA
#> 28               video/mp4       mp4              MPEG-4 video         -800
#> 29         application/pdf       pdf         Portable document           NA
#> 30               video/mp4       mp4              MPEG-4 video         -800
#> 31         application/pdf       pdf         Portable document           NA
#> 32               video/mp4       mp4              MPEG-4 video         -800
#> 33                text/csv       csv    Comma-separated values           NA
#> 34                text/csv       csv    Comma-separated values           NA
#> 35 application/vnd.datavyu       opf                   Datavyu           NA
#> 36         application/pdf       pdf         Portable document           NA
#> 37               image/png       png Portable network graphics           NA
#> 38               video/mp4       mp4              MPEG-4 video         -800
#> 39         application/pdf       pdf         Portable document           NA
#> 40         application/pdf       pdf         Portable document           NA
#> 41               video/mp4       mp4              MPEG-4 video         -800
#> 42         application/pdf       pdf         Portable document           NA
#> 43               video/mp4       mp4              MPEG-4 video         -800
#> 44               video/mp4       mp4              MPEG-4 video         -800
#> 45         application/pdf       pdf         Portable document           NA
#> 46               video/mp4       mp4              MPEG-4 video         -800
#> 47         application/pdf       pdf         Portable document           NA
#> 48         application/pdf       pdf         Portable document           NA
#> 49         application/pdf       pdf         Portable document           NA
#> 50               video/mp4       mp4              MPEG-4 video         -800
#> 51         application/pdf       pdf         Portable document           NA
#> 52               video/mp4       mp4              MPEG-4 video         -800
#> 53         application/pdf       pdf         Portable document           NA
#> 54               video/mp4       mp4              MPEG-4 video         -800
#> 55         application/pdf       pdf         Portable document           NA
#> 56         application/pdf       pdf         Portable document           NA
#> 57               video/mp4       mp4              MPEG-4 video         -800
#> 58         application/pdf       pdf         Portable document           NA
#> 59               video/mp4       mp4              MPEG-4 video         -800
#> 60         application/pdf       pdf         Portable document           NA
#> 61               video/mp4       mp4              MPEG-4 video         -800
#> 62         application/pdf       pdf         Portable document           NA
#> 63               video/mp4       mp4              MPEG-4 video         -800
#> 64         application/pdf       pdf         Portable document           NA
#> 65               video/mp4       mp4              MPEG-4 video         -800
#> 66         application/pdf       pdf         Portable document           NA
#> 67         application/pdf       pdf         Portable document           NA
#> 68         application/pdf       pdf         Portable document           NA
#> 69         application/pdf       pdf         Portable document           NA
#> 70         application/pdf       pdf         Portable document           NA
#> 71               video/mp4       mp4              MPEG-4 video         -800
#> 72         application/pdf       pdf         Portable document           NA
#> 73               video/mp4       mp4              MPEG-4 video         -800
#> 74         application/pdf       pdf         Portable document           NA
#> 75               video/mp4       mp4              MPEG-4 video         -800
#> 76         application/pdf       pdf         Portable document           NA
#> 77               video/mp4       mp4              MPEG-4 video         -800
#> 78         application/pdf       pdf         Portable document           NA
#> 79               video/mp4       mp4              MPEG-4 video         -800
#> 80               video/mp4       mp4              MPEG-4 video         -800
#> 81               video/mp4       mp4              MPEG-4 video         -800
#> 82         application/pdf       pdf         Portable document           NA
#> 83               video/mp4       mp4              MPEG-4 video         -800
#>    classification
#> 1              NA
#> 2              NA
#> 3              NA
#> 4              NA
#> 5              NA
#> 6               3
#> 7              NA
#> 8              NA
#> 9              NA
#> 10              3
#> 11             NA
#> 12             NA
#> 13             NA
#> 14             NA
#> 15             NA
#> 16             NA
#> 17             NA
#> 18             NA
#> 19             NA
#> 20             NA
#> 21             NA
#> 22             NA
#> 23             NA
#> 24             NA
#> 25             NA
#> 26             NA
#> 27              3
#> 28             NA
#> 29             NA
#> 30              3
#> 31             NA
#> 32             NA
#> 33             NA
#> 34              3
#> 35              3
#> 36              3
#> 37              3
#> 38              3
#> 39              3
#> 40             NA
#> 41             NA
#> 42             NA
#> 43             NA
#> 44             NA
#> 45             NA
#> 46             NA
#> 47              3
#> 48             NA
#> 49             NA
#> 50             NA
#> 51             NA
#> 52             NA
#> 53             NA
#> 54             NA
#> 55              3
#> 56             NA
#> 57             NA
#> 58             NA
#> 59             NA
#> 60             NA
#> 61             NA
#> 62             NA
#> 63             NA
#> 64              3
#> 65             NA
#> 66             NA
#> 67              3
#> 68              3
#> 69             NA
#> 70             NA
#> 71             NA
#> 72             NA
#> 73             NA
#> 74             NA
#> 75             NA
#> 76             NA
#> 77             NA
#> 78             NA
#> 79             NA
#> 80             NA
#> 81             NA
#> 82              3
#> 83             NA
#>  [ reached 'max' / getOption("max.print") -- omitted 45 rows ]