List Data For A Party (person or institution).

This function has been deprecated and may be removed in a future release. See get_party_by_id() for similar functionality.

list_party(party_id = 8, component = "all", vb = FALSE)

Arguments

party_id

Target volume number. Default is volume 8 (NYU).

component

Which data to return 'children', 'parents', or 'all'. Default is all.

vb

A Boolean value. If TRUE provides verbose output. Default is FALSE.

Value

A data frame with information about a party (person or institution).

Examples

# \donttest{
list_party() # Default is New York University (party 8)
#> Downloading: 3.9 kB     Downloading: 3.9 kB     Downloading: 3.9 kB     Downloading: 3.9 kB     
#> $id
#> [1] 8
#> 
#> $sortname
#> [1] "New York University (NYU)"
#> 
#> $url
#> [1] "http://www.nyu.edu/"
#> 
#> $institution
#> [1] TRUE
#> 
#> $parents
#>   party.id party.sortname party.institution
#> 1        0      Databrary              TRUE
#> 
#> $children
#>    party.id party.sortname party.prename             party.affiliation
#> 1         5         Adolph         Karen           New York University
#> 2         7        Millman         David           New York University
#> 3        11  Tamis-LeMonda     Catherine           New York University
#> 4       277         Rhodes      Marjorie           New York University
#> 5       335    Vouloumanos        Athena                          <NA>
#> 6       321         Marcus          Gary           New York University
#> 7       368          Raver        Cybele           New York University
#> 8      1098         Hughes         Diane           New York University
#> 9      3854          Merry   Sally Engle           New York University
#> 10     1254          Blair        Clancy           New York University
#> 11     3890          Sanes           Dan           New York University
#> 12     1834         Besser        Howard                           NYU
#> 13     2240      Yoshikawa      Hirokazu           New York University
#> 14     3802         Morris        Pamela           New York University
#> 15     3775             Ma       Jasmine           New York University
#> 16     5851        Picheny       Michael                           NYU
#> 17     2289    Schieffelin       Bambi B           New York University
#> 18     2355         Cheung        Olivia New York University Abu Dhabi
#> 19     2782      Gennetian          Lisa           New York University
#> 20     2810           West         Tessa                           NYU
#> 21     2936         Shrout    Patrick E.           New York University
#> 22     2953           Lake       Brenden           New York University
#> 23     3017         Dillon         Moira           New York University
#> 24     4685    Arunachalam         Sudha           New York University
#> 25     4790          Bloom        Stacie           New York University
#> 26     8726            Gao          Chen           New York University
#> 27     6517       Pargeter        Justin           New York University
#> 28     6507           Howe      Tsu-Hsin                          <NA>
#> 29     6499         Amodio         David           New York University
#> 30     6640        OConnor          Erin           New York University
#> 31     6697      Schneider         David           New York University
#> 32     6869        Trainor        Audrey           New York University
#> 33    11783         Fouhey         David           New York University
#> 34     9203       Ripolles         Pablo           New York University
#> 35     9752          Brito       Natalie           New York University
#> 36     9816         Grigos         Maria           New York University
#> 37     7136            Ren        Mengye           New York University
#> 38    11051          Pinto        Lerrel           New York University
#>                                party.url
#> 1       http://www.psych.nyu.edu/adolph/
#> 2                                   <NA>
#> 3                                   <NA>
#> 4                                   <NA>
#> 5  http://www.psych.nyu.edu/vouloumanos/
#> 6                                   <NA>
#> 7                                   <NA>
#> 8                                   <NA>
#> 9                                   <NA>
#> 10                                  <NA>
#> 11                                  <NA>
#> 12                                  <NA>
#> 13                                  <NA>
#> 14                                  <NA>
#> 15                                  <NA>
#> 16                                  <NA>
#> 17                                  <NA>
#> 18                                  <NA>
#> 19                                  <NA>
#> 20                                  <NA>
#> 21                                  <NA>
#> 22                                  <NA>
#> 23                                  <NA>
#> 24                                  <NA>
#> 25                                  <NA>
#> 26                                  <NA>
#> 27                                  <NA>
#> 28                                  <NA>
#> 29                                  <NA>
#> 30                                  <NA>
#> 31                                  <NA>
#> 32                                  <NA>
#> 33                                  <NA>
#> 34                                  <NA>
#> 35                                  <NA>
#> 36                                  <NA>
#> 37                 https://mengyeren.com
#> 38                                  <NA>
#> 
#> $access
#> list()
#> 
#' # }