List activity in a given Databrary session for which the user has ownership privileges.

List activity in a given Databrary session for which the user has ownership privileges.

Usage

list_session_activity(session_id = 6256, vb = FALSE)

Arguments

session_id

Selected session/slot number.

vb

A Boolean value. If TRUE provides verbose output.

Value

A list with the activity history on a slot.

Examples

list_session_activity()
#>    action              ip user       date                        when    id
#> 1     add 128.122.236.158    1 2013-10-28 2014-04-15T01:17:02.301985Z  6256
#> 2     add 128.122.236.158    1       <NA> 2014-04-15T01:17:06.754431Z  9826
#> 3     add 128.122.236.158    1       <NA> 2014-04-15T01:17:44.947757Z  9828
#> 4     add 128.122.236.158    1       <NA> 2014-04-15T01:18:23.294257Z  9830
#> 5     add 128.122.236.158    1       <NA> 2014-04-15T01:18:58.254206Z  9832
#> 6  change 128.122.236.158    1       <NA> 2014-04-15T01:25:58.481563Z    NA
#> 7     add 128.122.236.158    1       <NA> 2014-04-15T01:25:58.520207Z    NA
#> 8  change 128.122.236.158    1       <NA> 2014-04-15T01:27:14.660576Z    NA
#> 9     add  172.22.236.137    3       <NA> 2015-02-11T21:21:55.753231Z  9828
#> 10 remove  172.22.236.137    3       <NA> 2015-02-11T21:23:40.332373Z  9828
#> 11    add  172.22.236.137    3       <NA> 2015-02-11T21:26:26.717328Z  9828
#> 12 remove  172.22.236.137    3       <NA> 2015-02-11T21:29:36.508693Z  9828
#> 13    add  172.22.236.137    3       <NA> 2015-02-11T21:40:12.603296Z  9830
#> 14 remove  172.22.236.137    3       <NA> 2015-02-18T22:28:41.263953Z  9830
#> 15    add     100.2.20.44    3       <NA> 2015-03-04T23:16:58.898737Z  9830
#> 16 remove  172.22.236.137    3       <NA> 2015-04-01T16:33:59.256751Z  9830
#> 17 change  172.22.236.137    3       <NA> 2015-07-15T21:52:08.835224Z  9826
#> 18 change  172.22.236.137    3       <NA>  2015-07-15T21:52:53.77945Z  9828
#> 19 change  172.22.236.137    3       <NA> 2015-07-15T21:58:50.290747Z  9832
#> 20 change  172.22.236.137    3       <NA> 2015-07-15T21:59:13.210695Z  9830
#> 21    add  172.22.236.137    3       <NA> 2015-07-15T22:26:31.928748Z 22412
#> 22 change  172.22.236.137    3       <NA> 2015-07-15T22:32:03.799839Z  9828
#> 23 change 128.122.236.185   19       <NA> 2015-07-15T22:48:39.867839Z  9830
#> 24 change 128.122.236.185   19       <NA> 2015-07-15T22:51:35.057128Z  9830
#> 25 change 128.122.236.185   19       <NA> 2015-07-15T23:03:47.314683Z  9826
#> 26 change 128.122.236.148  293       <NA> 2015-08-13T21:26:39.634971Z  9826
#> 27 change 128.122.236.148  293       <NA> 2015-08-13T21:26:48.590024Z  9830
#>         type                              name           segment release
#> 1  container                              <NA>              NULL      NA
#> 2      asset                            Part 1         0, 335883      NA
#> 3      asset                            Part 2   335883, 4761366      NA
#> 4      asset                            Part 3  4761366, 9039201      NA
#> 5      asset                            Part 4 9039201, 12146348      NA
#> 6  container            Advisory Board Meeting              NULL      NA
#> 7    release                              <NA>              NULL       1
#> 8    release                              <NA>              NULL       3
#> 9    excerpt                              <NA>  2037802, 3019487      NA
#> 10   excerpt                              <NA>  2037802, 3019487      NA
#> 11   excerpt                              <NA>  2037802, 3019487      NA
#> 12   excerpt                              <NA>  2037802, 3019487      NA
#> 13   excerpt                              <NA>  5554168, 6616319      NA
#> 14   excerpt                              <NA>  5554168, 6616319      NA
#> 15   excerpt                              <NA>  6161659, 7146544      NA
#> 16   excerpt                              <NA>  6161659, 7146544      NA
#> 17     asset                      Introduction              NULL      NA
#> 18     asset       Overview and Databrary demo              NULL      NA
#> 19     asset                           Datavyu              NULL      NA
#> 20     asset               Databrary 1.0 plans              NULL      NA
#> 21     asset                            Slides              NULL      NA
#> 22     asset                    Databrary demo              NULL      NA
#> 23     asset          Databrary 1.0 plansdfdfd              NULL      NA
#> 24     asset Databrary 1.0 plansdfdfdsdfsdfsdf              NULL      NA
#> 25     asset                Introductionfdfhgh              NULL      NA
#> 26     asset                      Introduction              NULL      NA
#> 27     asset                Databrary 1.0 plan              NULL      NA
#>    old.release old.excerpt                                       old.name
#> 1           NA          NA                                           <NA>
#> 2           NA          NA                                           <NA>
#> 3           NA          NA                                           <NA>
#> 4           NA          NA                                           <NA>
#> 5           NA          NA                                           <NA>
#> 6           NA          NA                                           <NA>
#> 7           NA          NA                                           <NA>
#> 8            1          NA                                           <NA>
#> 9           NA          NA                                           <NA>
#> 10          NA           2                                           <NA>
#> 11          NA          NA                                           <NA>
#> 12          NA           2                                           <NA>
#> 13          NA          NA                                           <NA>
#> 14          NA          NA                                           <NA>
#> 15          NA          NA                                           <NA>
#> 16          NA          NA                                           <NA>
#> 17          NA          NA                                         Part 1
#> 18          NA          NA                                         Part 2
#> 19          NA          NA                                         Part 4
#> 20          NA          NA                                         Part 3
#> 21          NA          NA Databrary advisory board slides_2013-10-28.pdf
#> 22          NA          NA                    Overview and Databrary demo
#> 23          NA          NA                            Databrary 1.0 plans
#> 24          NA          NA                       Databrary 1.0 plansdfdfd
#> 25          NA          NA                                   Introduction
#> 26          NA          NA                             Introductionfdfhgh
#> 27          NA          NA              Databrary 1.0 plansdfdfdsdfsdfsdf
#>    excerpt
#> 1       NA
#> 2       NA
#> 3       NA
#> 4       NA
#> 5       NA
#> 6       NA
#> 7       NA
#> 8       NA
#> 9        2
#> 10      NA
#> 11       2
#> 12      NA
#> 13      NA
#> 14      NA
#> 15      NA
#> 16      NA
#> 17      NA
#> 18      NA
#> 19      NA
#> 20      NA
#> 21      NA
#> 22      NA
#> 23      NA
#> 24      NA
#> 25      NA
#> 26      NA
#> 27      NA