Get Stats About Databrary.

get_db_stats returns basic summary information about the institutions, people, and data hosted on 'Databrary.org'.

Usage

get_db_stats(type = "stats", vb = FALSE, rq = NULL)

Arguments

type

Type of Databrary report to run "institutions", "people", "data"

vb

A Boolean value. If TRUE provides verbose output.

rq

An httr2 request object.

Value

A data frame with the requested data or NULL if there is no new information.

Examples

# \donttest{
get_db_stats()
#> # A tibble: 1 × 9
#>   date                investigators affiliates institutions datasets_total
#>   <dttm>                      <int>      <int>        <int>          <int>
#> 1 2024-03-29 19:21:12          1740        680          784           1670
#> # ℹ 4 more variables: datasets_shared <int>, n_files <int>, hours <dbl>,
#> #   TB <dbl>
get_db_stats("stats")
#> # A tibble: 1 × 9
#>   date                investigators affiliates institutions datasets_total
#>   <dttm>                      <int>      <int>        <int>          <int>
#> 1 2024-03-29 19:21:12          1740        680          784           1670
#> # ℹ 4 more variables: datasets_shared <int>, n_files <int>, hours <dbl>,
#> #   TB <dbl>
get_db_stats("people") # Information about the newest authorized investigators.
#> # A tibble: 4 × 5
#>      id sortname prename     affiliation             time                       
#>   <int> <chr>    <chr>       <chr>                   <chr>                      
#> 1 12444 Naim     Reut        Tel-Aviv University     2024-03-28T14:35:50.569745Z
#> 2  1261 Oudeyer  Pierre-Yves Inria, France           2024-03-25T19:34:39.37165Z 
#> 3 12387 Camodeca Marina      University of Udine     2024-03-25T19:14:55.211561Z
#> 4 12346 Ledger   Susan       University of Newcastle 2024-03-25T15:48:02.372214Z
get_db_stats("places") # Information about the newest institutions.
#> # A tibble: 12 × 13
#>       id name       body  creation owners       permission publicsharefull time 
#>    <int> <chr>      <chr> <chr>    <list>            <int> <lgl>           <chr>
#>  1  1715 Training … "We … 2024-03… <named list>          1 FALSE           2024…
#>  2  1714 DBN - Uga… "Thi… 2024-03… <named list>          1 FALSE           2024…
#>  3  1713 DBN - Per… "Thi… 2024-03… <named list>          1 FALSE           2024…
#>  4 12444 NA          NA   NA       <NULL>               NA NA              2024…
#>  5  1712 Infant Sl… "Tas… 2024-03… <named list>          1 FALSE           2024…
#>  6  1711 Infant Sl… "Dat… 2024-03… <named list>          1 FALSE           2024…
#>  7  1261 NA          NA   NA       <NULL>               NA NA              2024…
#>  8 12508 NA          NA   NA       <NULL>               NA NA              2024…
#>  9 12387 NA          NA   NA       <NULL>               NA NA              2024…
#> 10 12507 NA          NA   NA       <NULL>               NA NA              2024…
#> 11  1710 Denver MA… "ful… 2024-03… <named list>          1 FALSE           2024…
#> 12 12346 NA          NA   NA       <NULL>               NA NA              2024…
#> # ℹ 5 more variables: sortname <chr>, prename <chr>, affiliation <chr>,
#> #   url <chr>, institution <lgl>
# }