List Assets in a Databrary Volume.

This function has been deprecated in favor of list_volume_assets().

list_assets_in_volume(vol_id = 2, vb = FALSE)

Arguments

vol_id

Target volume number.

vb

A boolean value.

Value

A data frame with information about all assets in a volume.

Examples

# \donttest{
list_assets_in_volume() # Assets in volume 2
#> Downloading: 990 B     Downloading: 990 B     Downloading: 990 B     Downloading: 990 B     Downloading: 810 B     Downloading: 810 B     Downloading: 810 B     Downloading: 810 B     Downloading: 800 B     Downloading: 800 B     Downloading: 800 B     Downloading: 800 B     Downloading: 290 B     Downloading: 290 B     Downloading: 290 B     Downloading: 290 B     Downloading: 240 B     Downloading: 240 B     Downloading: 240 B     Downloading: 240 B     
#>    asset_id asset_type_id duration          segment               name
#> 1         2          -800   633942        0, 633942        Rick part 1
#> 2         5          -800   634434        0, 634434     Florian part 1
#> 3         6          -800   380032  633906, 1013938     Florian part 2
#> 4        15          -800   378816  635094, 1013910        Rick part 2
#> 5        13          -800   177228 1013875, 1191103     Florian part 3
#> 6        12          -800   175979 1016187, 1192166        Rick part 3
#> 7        14          -800   252822        0, 252822     Florian part 1
#> 8         4          -800   253526        0, 253526        Rick part 1
#> 9         7          -800   217818   253531, 471349     Florian part 2
#> 10       11          -800   216832   255875, 472707        Rick part 2
#> 11       10          -800    67818   472687, 540505     Florian part 3
#> 12        8          -800    67222   474781, 542003        Rick part 3
#> 13        9          -800   662212        0, 662212            Florian
#> 14        3          -800   662102     1969, 664071               Rick
#> 15       17             2       NA             NULL 2013-8-30 metadata
#> 16       16             6       NA             NULL               Info
#>    permission      size        mimetype extension             asset_type
#> 1           1 499271255       video/mp4       mp4           MPEG-4 video
#> 2           1 466760824       video/mp4       mp4           MPEG-4 video
#> 3           1 296226810       video/mp4       mp4           MPEG-4 video
#> 4           1 360846586       video/mp4       mp4           MPEG-4 video
#> 5           1 126511616       video/mp4       mp4           MPEG-4 video
#> 6           1 151741089       video/mp4       mp4           MPEG-4 video
#> 7           1 192735602       video/mp4       mp4           MPEG-4 video
#> 8           1 218803157       video/mp4       mp4           MPEG-4 video
#> 9           1 148811524       video/mp4       mp4           MPEG-4 video
#> 10          1 173818370       video/mp4       mp4           MPEG-4 video
#> 11          1  26284689       video/mp4       mp4           MPEG-4 video
#> 12          1  22714940       video/mp4       mp4           MPEG-4 video
#> 13          1 470909386       video/mp4       mp4           MPEG-4 video
#> 14          1 416913623       video/mp4       mp4           MPEG-4 video
#> 15          1      1022        text/csv       csv Comma-separated values
#> 16          1     74400 application/pdf       pdf      Portable document
#>    transcodable classification
#> 1          -800             NA
#> 2          -800             NA
#> 3          -800             NA
#> 4          -800             NA
#> 5          -800             NA
#> 6          -800             NA
#> 7          -800             NA
#> 8          -800             NA
#> 9          -800             NA
#> 10         -800             NA
#> 11         -800             NA
#> 12         -800             NA
#> 13         -800             NA
#> 14         -800             NA
#> 15           NA              3
#> 16           NA              3
#' # }