Assets in Databrary Volume.

Assets in Databrary Volume.

Usage

list_assets_in_volume(vol_id = 1, vb = FALSE)

Arguments

vol_id

Target volume number.

vb

A boolean value.

Value

A data frame with information about all assets in a volume.

Examples

list_assets_in_volume() # Assets in volume 1
#>     asset_id asset_type_id
#> 1      22412             6
#> 2       9826          -800
#> 3       9828          -800
#> 4       9830          -800
#> 5       9832          -800
#> 6       9839             6
#> 7       9834          -800
#> 8       9836          -800
#> 9       9838          -800
#> 10     12039             6
#> 11     17737          -800
#> 12     22408             6
#> 13     17719          -800
#> 14     22407             6
#> 15     17723          -800
#> 16     17721          -800
#> 17     22410             6
#> 18     17725          -800
#> 19     22411             6
#> 20     17727          -800
#> 21     19919          -800
#> 22     22406             6
#> 23     25497             6
#> 24     25495             6
#> 25     25496             6
#> 26     27227             6
#> 27     26446             6
#> 28     28026          -800
#> 29     26449             6
#> 30     26448          -800
#> 31     26452             6
#> 32     28028          -800
#> 33    153109             2
#> 34    153108             2
#> 35    117035            16
#> 36       314             6
#> 37      9200             5
#> 38         1          -800
#> 39       315             6
#> 40     28047             6
#> 41     28051          -800
#> 42     28048             6
#> 43     28053          -800
#> 44     28055          -800
#> 45     28049             6
#> 46     28062          -800
#> 47     42470             6
#> 48     36978             6
#> 49     37128             6
#> 50     37138          -800
#> 51     36945             6
#> 52     37130          -800
#> 53     36946             6
#> 54     37136          -800
#> 55     43940             6
#> 56     38626             6
#> 57     38643          -800
#> 58     38627             6
#> 59     38591          -800
#> 60     38628             6
#> 61     38593          -800
#> 62     38629             6
#> 63     38595          -800
#> 64     43077             6
#> 65     44774          -800
#> 66     44544             6
#> 67     42740             6
#> 68     42755             6
#> 69     45413             6
#> 70     46104             6
#> 71     46300          -800
#> 72     60627             6
#> 73     60542          -800
#> 74     60628             6
#> 75     60544          -800
#> 76     60629             6
#> 77     60546          -800
#> 78     60630             6
#> 79     60550          -800
#> 80     61222          -800
#> 81     61224          -800
#> 82     65936             6
#> 83     66015          -800
#> 84     68596          -800
#> 85     68598          -800
#> 86     68600          -800
#> 87     68584          -800
#> 88     73302          -800
#> 89     73306          -800
#> 90     73304          -800
#> 91     73308          -800
#> 92     73310          -800
#> 93     73316          -800
#> 94     76809             6
#> 95     76808             6
#> 96     76807             6
#> 97     77809          -800
#> 98     77815          -800
#> 99     77811          -800
#> 100    77813          -800
#> 101    77817          -800
#> 102    77819          -800
#> 103    80993             6
#> 104    80894          -800
#> 105    81039          -600
#> 106    80896          -800
#> 107    80898          -800
#> 108    80950          -800
#> 109    79738          -800
#> 110    80952          -800
#> 111    80954          -800
#> 112    80958          -800
#> 113    80956          -800
#> 114    80960          -800
#> 115    80962          -800
#> 116    80964          -800
#> 117    84000          -800
#> 118    95780             6
#> 119   115695             6
#> 120   115751          -800
#> 121   115737          -600
#> 122   115747          -800
#> 123   115743          -800
#> 124   115745          -800
#> 125   115741          -800
#> 126   115753          -800
#> 127   115739          -800
#> 128   115749          -800
#>                                                                              name
#> 1                                                                          Slides
#> 2                                                                    Introduction
#> 3                                                                  Databrary demo
#> 4                                                              Databrary 1.0 plan
#> 5                                                                         Datavyu
#> 6                                                                    04-07-Slides
#> 7                                                      Overview and Policy Update
#> 8                                                               Private Beta demo
#> 9                                            Positioning Databrary for the future
#> 10                                                                         slides
#> 11                                                                  Board Meeting
#> 12                                                                 Welcome slides
#> 13                                                 Best practices in video coding
#> 14                                         SLIDES: Best practices in video coding
#> 15                                                               Datavyu tutorial
#> 16                                                               Lunch discussion
#> 17                                                       SLIDES: Lunch discussion
#> 18                                                        Open video data sharing
#> 19                                                SLIDES: Open video data sharing
#> 20                                                                 Databrary demo
#> 21                                                                          video
#> 22                                                                         slides
#> 23                                   Slides: Video reuse, management, and sharing
#> 24        Slides: Policies and best practices for sharing and managing video data
#> 25  Slides: Best practices in behavioral video coding and introduction to Datavyu
#> 26                                                                         Agenda
#> 27                                              SLIDES: Introduction to Databrary
#> 28                                                      Introduction to Databrary
#> 29            SLIDES: Policies & best practices for sharing & managing video data
#> 30                    Policies & best practices for sharing & managing video data
#> 31                 SLIDES: Introduction to Datavyu: Video coding & best practices
#> 32                         Introduction to Datavyu: Video coding & best practices
#> 33                                                      volumes-citations-monthly
#> 34                                           databrary-institutions-investigators
#> 35                                                   counting (demo Datavyu file)
#> 36                                                         NIH Proposal Narrative
#> 37                                                                           logo
#> 38                                                          counting (demo video)
#> 39                                                         NSF Proposal Narrative
#> 40                                              SLIDES: Introduction to Databrary
#> 41                                                      Introduction to Databrary
#> 42            SLIDES: Policies & best practices for sharing & managing video data
#> 43           Policies & best practices for sharing & managing video data (Part 1)
#> 44           Policies & best practices for sharing & managing video data (Part 2)
#> 45                 SLIDES: Introduction to Datavyu: Video coding & best practices
#> 46                         Introduction to Datavyu: Video coding & best practices
#> 47                                              Workshop attendees survey results
#> 48                              SLIDES: Datavyu & Best Practices for Coding Video
#> 49                                              SLIDES: Introduction to Databrary
#> 50                                                      Introduction to Databrary
#> 51                                SLIDES: Policies for Sharing Identifiable Video
#> 52                                        Policies for Sharing Identifiable Video
#> 53                                  SLIDES: Managing Data to Accelerate Discovery
#> 54                                          Managing Data to Accelerate Discovery
#> 55                                                                 Survey results
#> 56                                              SLIDES: Introduction to Databrary
#> 57                                                      Introduction to Databrary
#> 58                                SLIDES: Policies for Sharing Identifiable Video
#> 59                                        Policies for Sharing Identifiable Video
#> 60                                           SLIDES: Managing Data with Databrary
#> 61                                                   Managing Data with Databrary
#> 62                              SLIDES: Datavyu & Best Practices for Coding Video
#> 63                                      Datavyu & Best Practices for Coding Video
#> 64                                                            ICIS Survey Results
#> 65                                                       Databrary Pre-conference
#> 66                                              SLIDES: Introduction to Databrary
#> 67                                        SLIDES: Policies for Sharing Video Data
#> 68                                           SLIDES: Managing Data with Databrary
#> 69                              SLIDES: Datavyu & Best Practices for Coding Video
#> 70                                SLIDES: Databrary Board Meeting - June 13, 2016
#> 71                                 VIDEO: Databrary Board Meeting - June 13, 2016
#> 72                                              SLIDES: Introduction to Databrary
#> 73                                                      Introduction to Databrary
#> 74                                                   SLIDES: Policies for Sharing
#> 75                                                Policies for sharing video data
#> 76                                           SLIDES: Managing data with Databrary
#> 77                                                   Managing data with Databrary
#> 78                              SLIDES: Datavyu & best practices for coding video
#> 79                                      Datavyu & best Practices for coding video
#> 80                                                      Introduction to Databrary
#> 81                                                      Policies for sharing data
#> 82                                                PLAY workshop slides 2016-12-16
#> 83                                             PLAY workshop videocast 2016-12-16
#> 84                                                                            1/4
#> 85                                                                            2/4
#> 86                                                                            3/4
#> 87                                                                            4/4
#> 88                                                                          00002
#> 89                                                                          00003
#> 90                                                                          00004
#> 91                                                                          00005
#> 92                                                                          00006
#> 93                                     SRCD_2017_The_Science_of_Play_and_Learning
#> 94                                                      3-Datavyu-2017-04-30-aera
#> 95                                                 2-policiesMgmt-2017-04-30-aera
#> 96                                                        1-intro-2017-04-30-aera
#> 97                                                                          00012
#> 98                                                                          00013
#> 99                                                                          00014
#> 100                                                                        000151
#> 101                                                                        000152
#> 102                                                                         00016
#> 103                                Advisory Board Meeting Presentation_2017-06-09
#> 104                                                                         00019
#> 105         Heiman_Steinhardt Databrary_060917_GCASL CQ_PMO_Zoom4_Nus_bk# 1036791
#> 106                                                                         00020
#> 107                                                                         00021
#> 108                                                                         00022
#> 109                                  Advisory Board Meeting-2017-06-09 Part-2.mp4
#> 110                                                                         00023
#> 111                                                                         00024
#> 112                                                                         00026
#> 113                                                                         00025
#> 114                                                                         00027
#> 115                                                                         00028
#> 116                                                                         00029
#> 117                                             DatabraryAdvisoryBoardMeeting2017
#> 118                                    CDS_Databrary_Datavyu_2017-10-13_Final.key
#> 119                            Databrary_Advisory_Board_Meeting_slides-2018-03-26
#> 120                                             Advisory Board Meeting_2018_Part1
#> 121                                           Databrary Advisory Board_2018-03-26
#> 122                                             Advisory Board Meeting_2018_Part2
#> 123                                             Advisory Board Meeting_2018_Part3
#> 124                                             Advisory Board Meeting_2018_Part4
#> 125                                             Advisory Board Meeting_2018_Part5
#> 126                                             Advisory Board Meeting_2018_Part6
#> 127                                             Advisory Board Meeting_2018_Part7
#> 128                                             Advisory Board Meeting_2018_Part8
#>     permission       size duration            segment                mimetype
#> 1            1    4573426       NA               NULL         application/pdf
#> 2            1   88610655   335883          0, 335883               video/mp4
#> 3            1  917124852  4425483    335883, 4761366               video/mp4
#> 4            1  899912341  4277835   4761366, 9039201               video/mp4
#> 5            1  764340542  3107147  9039201, 12146348               video/mp4
#> 6            1    3013342       NA               NULL         application/pdf
#> 7            1 1301079971  4964011         0, 4964011               video/mp4
#> 8            1 1419151887  5043051  4964011, 10007062               video/mp4
#> 9            1  882417129  3821387 10007062, 13828449               video/mp4
#> 10           1   10192940       NA               NULL         application/pdf
#> 11           1 5316455384 10976384        0, 10976384               video/mp4
#> 12           1     304566       NA                  0         application/pdf
#> 13           1 1384301751  3514603         0, 3514603               video/mp4
#> 14           1    2924252       NA             528181         application/pdf
#> 15           1  574598972  1594688   3514560, 5109248               video/mp4
#> 16           1 1602298195  3977558   5111054, 9088612               video/mp4
#> 17           1    6899381       NA            5112000         application/pdf
#> 18           1  710384623  1920512  9086719, 11007231               video/mp4
#> 19           1    4158160       NA            9087000         application/pdf
#> 20           1 1079191541  2827904 11007188, 13835092               video/mp4
#> 21           1 5561649547 14939008        0, 14939008               video/mp4
#> 22           1   24014506       NA               5000         application/pdf
#> 23           1   19543511       NA                  0         application/pdf
#> 24           1    6947725       NA               1000         application/pdf
#> 25           1   11216184       NA               2000         application/pdf
#> 26           1      55248       NA                  0         application/pdf
#> 27           1    8502489       NA                100         application/pdf
#> 28           1  969574682  3708160   1061959, 4770119               video/mp4
#> 29           1   18074689       NA            4741737         application/pdf
#> 30           1 2059896165  5477483  4742737, 10220220               video/mp4
#> 31           1   16562761       NA           10220209         application/pdf
#> 32           1  717531830  2725723 10220220, 12945943               video/mp4
#> 33           0         NA       NA               NULL                text/csv
#> 34           1       1826       NA               NULL                text/csv
#> 35           1       1030       NA               NULL application/vnd.datavyu
#> 36           1    3826718       NA               NULL         application/pdf
#> 37           1     104419       NA               NULL               image/png
#> 38           1     394499    40000               NULL               video/mp4
#> 39           1    4120930       NA               NULL         application/pdf
#> 40           1    9500786       NA                  0         application/pdf
#> 41           1  886710056  2794880         0, 2794880               video/mp4
#> 42           1   18069955       NA            2792054         application/pdf
#> 43           1  763350644  2387051   2795785, 5182836               video/mp4
#> 44           1 1044753106  2983616   5183754, 8167370               video/mp4
#> 45           1   17821836       NA            8150330         application/pdf
#> 46           1 1550955220  4835115  8168370, 13003485               video/mp4
#> 47           1     402918       NA               NULL         application/pdf
#> 48           1   16143550       NA               NULL         application/pdf
#> 49           1   22240020       NA                  0         application/pdf
#> 50           1 1190956598  3402326         0, 3402326               video/mp4
#> 51           1    2225737       NA            1800000         application/pdf
#> 52           1  824306229  2450882   1827192, 4278074               video/mp4
#> 53           1   13901204       NA            2400000         application/pdf
#> 54           1 1412799175  3904128   2451165, 6355293               video/mp4
#> 55           1     191095       NA               NULL         application/pdf
#> 56           1   17971722       NA                  0         application/pdf
#> 57           1  706313005  2760992         0, 2760992               video/mp4
#> 58           1    4265477       NA            2811000         application/pdf
#> 59           1  622855933  2293056   2814628, 5107684               video/mp4
#> 60           1   10040436       NA            5095000         application/pdf
#> 61           1 1035603796  3676480   5108605, 8785085               video/mp4
#> 62           1   11171502       NA            8784000         application/pdf
#> 63           1  857979902  2537494  8785992, 11323486               video/mp4
#> 64           1    1772791       NA               NULL         application/pdf
#> 65           1 5207302273  9728726         0, 9728726               video/mp4
#> 66           1   10732100       NA               2000         application/pdf
#> 67           1    4238406       NA            2185146         application/pdf
#> 68           1    7741745       NA            4401856         application/pdf
#> 69           1   15475455       NA            6480140         application/pdf
#> 70           1   19602134       NA               NULL         application/pdf
#> 71           1 4651981876 13129280        0, 13129280               video/mp4
#> 72           1   10732100       NA                  0         application/pdf
#> 73           1  607629550  2058198         0, 2058198               video/mp4
#> 74           1    4238406       NA            2056000         application/pdf
#> 75           1  795931994  2468587   2059089, 4527676               video/mp4
#> 76           1    7741745       NA            4524000         application/pdf
#> 77           1  627556299  2002624   4528588, 6531212               video/mp4
#> 78           1   15475455       NA            6527000         application/pdf
#> 79           1 1033985826  3372779   6532121, 9904900               video/mp4
#> 80           1  671824312  1939563         0, 1939563               video/mp4
#> 81           1  687136253  1908544   1940470, 3849014               video/mp4
#> 82           1  278432416       NA               NULL         application/pdf
#> 83           1 1332848036 24567644        0, 24567644               video/mp4
#> 84           1 1251324058  2499595         0, 2499595               video/mp4
#> 85           1 1322485833  2857388   2500552, 5357940               video/mp4
#> 86           1 1310253381  2857722   5357939, 8215661               video/mp4
#> 87           1  984178185  2096523  8215657, 10312180               video/mp4
#> 88           1 1141378800  2853384         0, 2853384               video/mp4
#> 89           1 1109757036  2859724   2854383, 5714107               video/mp4
#> 90           1  308336133   798699   5714099, 6512798               video/mp4
#> 91           1 1077537629  2856387   6513755, 9370142               video/mp4
#> 92           1  555319245  1435851  9371141, 10806992               video/mp4
#> 93           1 1185980540  2855887         0, 2855887               video/mp4
#> 94           1   11296910       NA               NULL         application/pdf
#> 95           1    8188127       NA               NULL         application/pdf
#> 96           1    9899927       NA               NULL         application/pdf
#> 97           1  977096783  2853384         0, 2853384               video/mp4
#> 98           1  943655212  2855220   2854383, 5709603               video/mp4
#> 99           1   39116816   121547   5709594, 5831141               video/mp4
#> 100          1  243847194   704538   5831098, 6535636               video/mp4
#> 101          1  695084769  2037970   6535606, 8573576               video/mp4
#> 102          1  252821639   743179   8573092, 9316271               video/mp4
#> 103          1   81457940       NA               NULL         application/pdf
#> 104          1 1185269159  1366382         0, 1366382               video/mp4
#> 105          1  292650423 15823072        0, 15823072              audio/mpeg
#> 106          1 1244239760  1368451   1366381, 2734832               video/mp4
#> 107          1 1191221405  1366949   2734837, 4101786               video/mp4
#> 108          1 1237461468  1346475   4101786, 5448261               video/mp4
#> 109          1   86403462  4950334   4832834, 9783168               video/mp4
#> 110          1 1113502210  1366382   5448218, 6814600               video/mp4
#> 111          1 1040993535  1367450   6814599, 8182049               video/mp4
#> 112          1  946159516  1101739   8182054, 9283793               video/mp4
#> 113          1 1111936525  1367934  9283750, 10651684               video/mp4
#> 114          1 1074824135  1374390 10651691, 12026081               video/mp4
#> 115          1 1093669899  1375775 12026080, 13401855               video/mp4
#> 116          1 1029314077  1251691 13401859, 14653550               video/mp4
#> 117          1 4611531493 14589934 15824072, 30414006               video/mp4
#> 118          1   24630245       NA               NULL         application/pdf
#> 119          1   53661203       NA               NULL         application/pdf
#> 120          1 1886564537  2179711         0, 2179711               video/mp4
#> 121          1  226542146 14398328        0, 14398328              audio/mpeg
#> 122          1 1976644381  2181547   2179694, 4361241               video/mp4
#> 123          1  581869073   602027   4361207, 4963234               video/mp4
#> 124          1 2321264233  2180712   4802729, 6983441               video/mp4
#> 125          1 2084960359  2180546   6983424, 9163970               video/mp4
#> 126          1  184972983   159595   9163945, 9323540               video/mp4
#> 127          1 1985795461  2180212  9323497, 11503709               video/mp4
#> 128          1  554267042   504427 11503691, 12008118               video/mp4
#>     extension                       asset_type transcodable classification
#> 1         pdf                Portable document           NA             NA
#> 2         mp4                     MPEG-4 video         -800             NA
#> 3         mp4                     MPEG-4 video         -800             NA
#> 4         mp4                     MPEG-4 video         -800             NA
#> 5         mp4                     MPEG-4 video         -800             NA
#> 6         pdf                Portable document           NA              3
#> 7         mp4                     MPEG-4 video         -800             NA
#> 8         mp4                     MPEG-4 video         -800             NA
#> 9         mp4                     MPEG-4 video         -800             NA
#> 10        pdf                Portable document           NA              3
#> 11        mp4                     MPEG-4 video         -800             NA
#> 12        pdf                Portable document           NA             NA
#> 13        mp4                     MPEG-4 video         -800             NA
#> 14        pdf                Portable document           NA             NA
#> 15        mp4                     MPEG-4 video         -800             NA
#> 16        mp4                     MPEG-4 video         -800             NA
#> 17        pdf                Portable document           NA             NA
#> 18        mp4                     MPEG-4 video         -800             NA
#> 19        pdf                Portable document           NA             NA
#> 20        mp4                     MPEG-4 video         -800             NA
#> 21        mp4                     MPEG-4 video         -800             NA
#> 22        pdf                Portable document           NA             NA
#> 23        pdf                Portable document           NA             NA
#> 24        pdf                Portable document           NA             NA
#> 25        pdf                Portable document           NA             NA
#> 26        pdf                Portable document           NA             NA
#> 27        pdf                Portable document           NA              3
#> 28        mp4                     MPEG-4 video         -800             NA
#> 29        pdf                Portable document           NA             NA
#> 30        mp4                     MPEG-4 video         -800              3
#> 31        pdf                Portable document           NA             NA
#> 32        mp4                     MPEG-4 video         -800             NA
#> 33        csv           Comma-separated values           NA             NA
#> 34        csv           Comma-separated values           NA              3
#> 35        opf                          Datavyu           NA              3
#> 36        pdf                Portable document           NA              3
#> 37        png        Portable network graphics           NA              3
#> 38        mp4                     MPEG-4 video         -800              3
#> 39        pdf                Portable document           NA              3
#> 40        pdf                Portable document           NA             NA
#> 41        mp4                     MPEG-4 video         -800             NA
#> 42        pdf                Portable document           NA             NA
#> 43        mp4                     MPEG-4 video         -800             NA
#> 44        mp4                     MPEG-4 video         -800             NA
#> 45        pdf                Portable document           NA             NA
#> 46        mp4                     MPEG-4 video         -800             NA
#> 47        pdf                Portable document           NA              3
#> 48        pdf                Portable document           NA             NA
#> 49        pdf                Portable document           NA             NA
#> 50        mp4                     MPEG-4 video         -800             NA
#> 51        pdf                Portable document           NA             NA
#> 52        mp4                     MPEG-4 video         -800             NA
#> 53        pdf                Portable document           NA             NA
#> 54        mp4                     MPEG-4 video         -800             NA
#> 55        pdf                Portable document           NA              3
#> 56        pdf                Portable document           NA             NA
#> 57        mp4                     MPEG-4 video         -800             NA
#> 58        pdf                Portable document           NA             NA
#> 59        mp4                     MPEG-4 video         -800             NA
#> 60        pdf                Portable document           NA             NA
#> 61        mp4                     MPEG-4 video         -800             NA
#> 62        pdf                Portable document           NA             NA
#> 63        mp4                     MPEG-4 video         -800             NA
#> 64        pdf                Portable document           NA              3
#> 65        mp4                     MPEG-4 video         -800             NA
#> 66        pdf                Portable document           NA             NA
#> 67        pdf                Portable document           NA              3
#> 68        pdf                Portable document           NA              3
#> 69        pdf                Portable document           NA             NA
#> 70        pdf                Portable document           NA             NA
#> 71        mp4                     MPEG-4 video         -800             NA
#> 72        pdf                Portable document           NA             NA
#> 73        mp4                     MPEG-4 video         -800             NA
#> 74        pdf                Portable document           NA             NA
#> 75        mp4                     MPEG-4 video         -800             NA
#> 76        pdf                Portable document           NA             NA
#> 77        mp4                     MPEG-4 video         -800             NA
#> 78        pdf                Portable document           NA             NA
#> 79        mp4                     MPEG-4 video         -800             NA
#> 80        mp4                     MPEG-4 video         -800             NA
#> 81        mp4                     MPEG-4 video         -800             NA
#> 82        pdf                Portable document           NA              3
#> 83        mp4                     MPEG-4 video         -800             NA
#> 84        mp4                     MPEG-4 video         -800              3
#> 85        mp4                     MPEG-4 video         -800             NA
#> 86        mp4                     MPEG-4 video         -800             NA
#> 87        mp4                     MPEG-4 video         -800             NA
#> 88        mp4                     MPEG-4 video         -800             NA
#> 89        mp4                     MPEG-4 video         -800             NA
#> 90        mp4                     MPEG-4 video         -800             NA
#> 91        mp4                     MPEG-4 video         -800             NA
#> 92        mp4                     MPEG-4 video         -800             NA
#> 93        mp4                     MPEG-4 video         -800             NA
#> 94        pdf                Portable document           NA             NA
#> 95        pdf                Portable document           NA             NA
#> 96        pdf                Portable document           NA             NA
#> 97        mp4                     MPEG-4 video         -800             NA
#> 98        mp4                     MPEG-4 video         -800             NA
#> 99        mp4                     MPEG-4 video         -800             NA
#> 100       mp4                     MPEG-4 video         -800             NA
#> 101       mp4                     MPEG-4 video         -800             NA
#> 102       mp4                     MPEG-4 video         -800             NA
#> 103       pdf                Portable document           NA             NA
#> 104       mp4                     MPEG-4 video         -800             NA
#> 105       mp3 MPEG-1 or MPEG-2 audio layer III         -600             NA
#> 106       mp4                     MPEG-4 video         -800             NA
#> 107       mp4                     MPEG-4 video         -800             NA
#> 108       mp4                     MPEG-4 video         -800             NA
#> 109       mp4                     MPEG-4 video         -800              3
#> 110       mp4                     MPEG-4 video         -800             NA
#> 111       mp4                     MPEG-4 video         -800             NA
#> 112       mp4                     MPEG-4 video         -800             NA
#> 113       mp4                     MPEG-4 video         -800             NA
#> 114       mp4                     MPEG-4 video         -800             NA
#> 115       mp4                     MPEG-4 video         -800             NA
#> 116       mp4                     MPEG-4 video         -800             NA
#> 117       mp4                     MPEG-4 video         -800             NA
#> 118       pdf                Portable document           NA             NA
#> 119       pdf                Portable document           NA             NA
#> 120       mp4                     MPEG-4 video         -800             NA
#> 121       mp3 MPEG-1 or MPEG-2 audio layer III         -600             NA
#> 122       mp4                     MPEG-4 video         -800             NA
#> 123       mp4                     MPEG-4 video         -800             NA
#> 124       mp4                     MPEG-4 video         -800             NA
#> 125       mp4                     MPEG-4 video         -800             NA
#> 126       mp4                     MPEG-4 video         -800             NA
#> 127       mp4                     MPEG-4 video         -800             NA
#> 128       mp4                     MPEG-4 video         -800             NA