List Image or Video Excerpts On A Databrary Volume.

List Image or Video Excerpts On A Databrary Volume.

Usage

list_volume_excerpts(vol_id = 1, vb = options::opt("vb"), rq = NULL)

Arguments

vol_id

Target volume number.

vb

Show verbose messages. (Defaults to FALSE, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')

rq

An httr2 request object. Default is NULL.

Value

A list with information about any available excerpts.

Examples

# \donttest{
list_volume_excerpts()
#> $id
#> [1] 1
#> 
#> $name
#> [1] "Databrary sponsored workshops and events"
#> 
#> $body
#> [1] "Our mission is to promote sharing and reuse of video data. Most developmental scientists rely on video recordings to capture the complexity and richness of behavior. However, researchers rarely share video data, and this has impeded scientific progress. Enabling reuse of video data is essential to increasing openness, maximizing the efficiency of limited resources, and accelerating the pace of discovery in the developmental and learning sciences.\n\nWe aim to provide tools that encourage good data management, facilitate open data sharing, enable discovery, and foster data reuse and integration. We are working to build functionality that allows researchers to organize, store, code, and share identifiable video data and other contextual information about the videos with other researchers through a secure web-based library.\n\nThe Databrary project is dedicated to transforming the culture of developmental science by building a community of researchers committed to open video data sharing, training a new generation of developmental scientists and empowering them with an unprecedented set of tools for discovery, and raising the profile of behavioral science by bolstering interest in and support for scientific research among the general public."
#> 
#> $doi
#> [1] "10.17910/B7159Q"
#> 
#> $creation
#> [1] "2013-01-11T10:26:40Z"
#> 
#> $owners
#> $owners[[1]]
#> $owners[[1]]$name
#> [1] "Adolph, Karen"
#> 
#> $owners[[1]]$id
#> [1] 5
#> 
#> 
#> $owners[[2]]
#> $owners[[2]]$name
#> [1] "Gilmore, Rick O."
#> 
#> $owners[[2]]$id
#> [1] 6
#> 
#> 
#> $owners[[3]]
#> $owners[[3]]$name
#> [1] "Staff"
#> 
#> $owners[[3]]$id
#> [1] 308
#> 
#> 
#> $owners[[4]]
#> $owners[[4]]$name
#> [1] "Admin, Databrary"
#> 
#> $owners[[4]]$id
#> [1] 3341
#> 
#> 
#> 
#> $permission
#> [1] 1
#> 
#> $publicsharefull
#> [1] TRUE
#> 
#> $publicaccess
#> [1] "full"
#> 
#> $excerpts
#> $excerpts[[1]]
#> $excerpts[[1]]$segment
#> $excerpts[[1]]$segment[[1]]
#> [1] 16852000
#> 
#> $excerpts[[1]]$segment[[2]]
#> [1] 16916503
#> 
#> 
#> $excerpts[[1]]$permission
#> [1] 1
#> 
#> $excerpts[[1]]$excerpt
#> NULL
#> 
#> $excerpts[[1]]$asset
#> $excerpts[[1]]$asset$id
#> [1] 84000
#> 
#> $excerpts[[1]]$asset$format
#> [1] -800
#> 
#> $excerpts[[1]]$asset$duration
#> [1] 14589934
#> 
#> $excerpts[[1]]$asset$segment
#> $excerpts[[1]]$asset$segment[[1]]
#> [1] 15824072
#> 
#> $excerpts[[1]]$asset$segment[[2]]
#> [1] 30414006
#> 
#> 
#> $excerpts[[1]]$asset$name
#> [1] "DatabraryAdvisoryBoardMeeting2017"
#> 
#> $excerpts[[1]]$asset$permission
#> [1] 1
#> 
#> $excerpts[[1]]$asset$size
#> [1] 4611531493
#> 
#> $excerpts[[1]]$asset$container
#> [1] 17886
#> 
#> 
#> 
#> $excerpts[[2]]
#> $excerpts[[2]]$segment
#> [1] 0
#> 
#> $excerpts[[2]]$permission
#> [1] 1
#> 
#> $excerpts[[2]]$excerpt
#> NULL
#> 
#> $excerpts[[2]]$asset
#> $excerpts[[2]]$asset$id
#> [1] 9200
#> 
#> $excerpts[[2]]$asset$format
#> [1] 5
#> 
#> $excerpts[[2]]$asset$classification
#> [1] 3
#> 
#> $excerpts[[2]]$asset$name
#> [1] "logo"
#> 
#> $excerpts[[2]]$asset$permission
#> [1] 1
#> 
#> $excerpts[[2]]$asset$size
#> [1] 104419
#> 
#> $excerpts[[2]]$asset$container
#> [1] 9807
#> 
#> 
#> 
#> 
# }